Split linux/macos unmount impl
