+++ src/storage/nix/linux.rs
@@ -122,3 +122,13 @@ fn build_device_info(d: &Block) -> Option<StorageRef> {
         children,
     })
 }
+
+pub(crate) fn unmount(path: &str) -> Result<()> {
+    let cpath = CString::new(path)?;
+    match unsafe { libc::umount2(cpath.as_ptr(), libc::MNT_FORCE) } {
+        0 => Ok(()),
+        _ if std::io::Error::last_os_error().raw_os_error() == Some(libc::ENOENT) => Ok(()), // not found
+        _ => Err(anyhow::Error::new(std::io::Error::last_os_error())
+            .context("Failed to unmount a volume")),
+    }
+}
