+++ src/storage/nix/mod.rs
@@ -19,7 +19,6 @@ mod macos;
 
 #[cfg(target_os = "macos")]
 use macos as os;
-use std::ffi::CString;
 
 impl StorageError {
     fn from(err: std::io::Error) -> StorageError {
@@ -85,25 +84,14 @@ impl StorageDevice for StorageRef {
             .iter()
             .flat_map(|c| &c.details.mount_point)
             .for_each(|c| {
-                let _ = unmount(c.as_str());
+                let _ = os::unmount(c.as_str());
             });
 
         match &self.details.mount_point {
-            Some(c) => unmount(c.as_str())?,
+            Some(c) => os::unmount(c.as_str())?,
             _ => (),
         };
 
         FileAccess::new(&self.id).map(|a| Box::new(a) as Box<dyn StorageAccess>)
-    }
-}
-
-fn unmount(path: &str) -> Result<()> {
-    let cpath = CString::new(path)?;
-    match unsafe { libc::unmount(cpath.as_ptr(), libc::MNT_FORCE) } {
-        0 => Ok(()),
-        _ if std::io::Error::last_os_error().raw_os_error() == Some(libc::ENOENT) => Ok(()), // not found
-        _ => Err(anyhow::Error::new(
-            std::io::Error::last_os_error()
-        ).context("Failed to unmount a volume")),
     }
 }
